test(decode): style test file
