+++ src/decode.rs
@@ -105,6 +105,9 @@ use crate::{
     types::object::{InvalidObjectDir, InvalidObjectPath},
 };
 
+#[cfg(test)]
+mod test;
+
 /// 将一个 object 的数据写入到 rust 类型
 pub trait RefineObject<Error: StdError + 'static> {
     /// 提取 key
@@ -661,227 +664,4 @@ pub enum ListErrorKind {
 
     #[non_exhaustive]
     Custom(Box<dyn StdError + 'static>),
-}
-
-#[cfg(test)]
-mod tests {
-
-    use std::fmt;
-
-    use super::*;
-
-    #[test]
-    fn test_one_object_decode() {
-        struct ObjectA {}
-        impl RefineObject<MyError> for ObjectA {
-            fn set_key(&mut self, key: &str) -> Result<(), MyError> {
-                assert!(key == "LICENSE");
-                Ok(())
-            }
-            fn set_last_modified(&mut self, last_modified: &str) -> Result<(), MyError> {
-                assert!(last_modified == "2022-06-12T06:11:06.000Z");
-                Ok(())
-            }
-            fn set_etag(&mut self, etag: &str) -> Result<(), MyError> {
-                assert!(etag == "2CBAB10A50CC6905EA2D7CCCEF31A6C9");
-                Ok(())
-            }
-            fn set_type(&mut self, _type: &str) -> Result<(), MyError> {
-                assert!(_type == "Normal");
-                Ok(())
-            }
-            fn set_size(&mut self, size: &str) -> Result<(), MyError> {
-                assert!(size == "1065");
-                Ok(())
-            }
-            fn set_storage_class(&mut self, storage_class: &str) -> Result<(), MyError> {
-                assert!(storage_class == "Standard");
-                Ok(())
-            }
-        }
-
-        #[derive(Debug)]
-        struct MyError {}
-
-        impl From<InnerItemError> for MyError {
-            fn from(_: InnerItemError) -> Self {
-                MyError {}
-            }
-        }
-
-        impl From<quick_xml::Error> for MyError {
-            fn from(_: quick_xml::Error) -> Self {
-                MyError {}
-            }
-        }
-
-        impl fmt::Display for MyError {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                write!(f, "demo")
-            }
-        }
-        impl std::error::Error for MyError {}
-
-        let xml = r#"<Key>LICENSE</Key>
-            <LastModified>2022-06-12T06:11:06.000Z</LastModified>
-            <ETag>"2CBAB10A50CC6905EA2D7CCCEF31A6C9"</ETag>
-            <Type>Normal</Type>
-            <Size>1065</Size>
-            <StorageClass>Standard</StorageClass>"#;
-
-        let mut object = ObjectA {};
-        let _ = object.decode(xml);
-    }
-
-    #[test]
-    fn test_common_prefixes() {
-        use std::error::Error;
-        struct ObjectA {}
-        impl RefineObject<MyError> for ObjectA {}
-
-        struct ListA {}
-
-        #[derive(Debug)]
-        struct MyError {}
-
-        impl From<InnerItemError> for MyError {
-            fn from(_: InnerItemError) -> Self {
-                MyError {}
-            }
-        }
-
-        impl From<quick_xml::Error> for MyError {
-            fn from(_: quick_xml::Error) -> Self {
-                MyError {}
-            }
-        }
-
-        impl fmt::Display for MyError {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                write!(f, "demo")
-            }
-        }
-
-        impl Error for MyError {}
-        impl ListError for MyError {}
-
-        impl RefineObjectList<ObjectA, MyError, MyError> for ListA {
-            fn set_prefix(&mut self, prefix: &str) -> Result<(), MyError> {
-                assert!(prefix == "bar");
-                Ok(())
-            }
-
-            fn set_common_prefix(&mut self, list: &[Cow<'_, str>]) -> Result<(), MyError> {
-                assert!(list[0] == "foo1/");
-                assert!(list[1] == "foo2/");
-                Ok(())
-            }
-        }
-
-        let xml = r#"<?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult>
-          <Prefix>bar</Prefix>
-          <Contents>
-            <Key>9AB932LY.jpeg</Key>
-          </Contents>
-          <Contents>
-            <Key>9AB932LY.jpeg</Key>
-          </Contents>
-          <CommonPrefixes>
-            <Prefix>foo1/</Prefix>
-            <Prefix>foo2/</Prefix>
-          </CommonPrefixes>
-        </ListBucketResult>
-        "#;
-
-        let mut list = ListA {};
-
-        let init_object = || ObjectA {};
-
-        let res = list.decode(xml, init_object);
-
-        assert!(res.is_ok());
-    }
-
-    #[test]
-    fn test_item_from() {
-        let err = InnerItemError::new();
-        assert_eq!(format!("{err}"), "demo");
-    }
-
-    #[test]
-    fn test_error_item_source() {
-        let err = InnerItemError::new();
-        assert!(err.get_source().is_none());
-    }
-
-    #[test]
-    fn test_list_from() {
-        let string = InvalidEndPoint { _priv: () };
-        let err: InnerListError = string.into();
-        assert_eq!(
-            format!("{err}"),
-            "endpoint must not with `-` prefix or `-` suffix or `oss-` prefix"
-        );
-    }
-
-    #[test]
-    fn test_error_list_display() {
-        let err = InnerItemError::new();
-        let err_list: InnerListError = err.into();
-        assert_eq!(err_list.to_string(), "demo");
-
-        let err = InnerListError::from_xml();
-        assert_eq!(format!("{err}"), "Cannot read text, expecting Event::Text");
-
-        let err = InnerListError::from_custom();
-        assert_eq!(format!("{err}"), "custom");
-    }
-
-    #[test]
-    fn test_error_list_get_source() {
-        let err = InnerItemError::new();
-        let err_list: InnerListError = err.into();
-        assert!(err_list.get_source().is_none());
-
-        let err_list = InnerListError::from_xml();
-        assert!(err_list.get_source().is_none());
-
-        let err_list = InnerListError::from_custom();
-        assert!(err_list.get_source().is_none());
-    }
-
-    #[test]
-    fn test_error_list_from_item() {
-        let err = InnerListError {
-            kind: ListErrorKind::Xml(quick_xml::Error::TextNotFound),
-        };
-        assert_eq!(format!("{err}"), "Cannot read text, expecting Event::Text");
-
-        fn bar() -> InnerListError {
-            InnerItemError::new().into()
-        }
-
-        assert_eq!(
-            format!("{:?}", bar()),
-            "InnerListError { kind: Item(InnerItemError(MyError)) }"
-        );
-    }
-
-    #[test]
-    fn test_error_list_from_xml() {
-        let err = InnerListError {
-            kind: ListErrorKind::Xml(quick_xml::Error::TextNotFound),
-        };
-        assert_eq!(format!("{err}"), "Cannot read text, expecting Event::Text");
-
-        fn bar() -> InnerListError {
-            quick_xml::Error::TextNotFound.into()
-        }
-
-        assert_eq!(
-            format!("{:?}", bar()),
-            "InnerListError { kind: Xml(TextNotFound) }"
-        );
-    }
 }
