+++ src/tests/mod.rs
@@ -1,9 +1,6 @@
 #[cfg(feature = "core")]
 mod client;
 
-#[cfg(feature = "decode")]
-mod traits;
-
 #[cfg(feature = "core")]
 mod errors;
